+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatIPWindow.cs
@@ -1,8 +1,5 @@
-#if PLATFORM_ANDROID
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
-using UnityEditor.Android;
 using System.Text.RegularExpressions;
 
 namespace Unity.Android.Logcat
@@ -30,6 +27,8 @@ namespace Unity.Android.Logcat
 
         void OnEnable()
         {
+            if (!AndroidBridge.AndroidExtensionsInstalled)
+                return;
             if (m_Runtime == null)
                 m_Runtime = AndroidLogcatManager.instance.Runtime;
             m_IpString = EditorPrefs.GetString(kAndroidLogcatLastIp, "");
@@ -43,6 +42,8 @@ namespace Unity.Android.Logcat
 
         private void OnDisable()
         {
+            if (!AndroidBridge.AndroidExtensionsInstalled)
+                return;
             if (m_Runtime == null)
                 return;
             m_Runtime.DeviceQuery.DevicesUpdated -= DevicesUpdated;
@@ -110,6 +111,12 @@ namespace Unity.Android.Logcat
 
         void OnGUI()
         {
+            if (!AndroidBridge.AndroidExtensionsInstalled)
+            {
+                AndroidLogcatUtilities.ShowAndroidIsNotInstalledMessage();
+                return;
+            }
+
             EditorGUILayout.BeginVertical();
             {
                 EditorGUILayout.LabelField("Available devices:", EditorStyles.boldLabel);
@@ -203,4 +210,3 @@ namespace Unity.Android.Logcat
         }
     }
 }
-#endif