+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatManager.cs
@@ -1,13 +1,4 @@
-#if PLATFORM_ANDROID
-using System.Collections.Generic;
-using System.Diagnostics;
-using System;
-using System.Text.RegularExpressions;
-using System.Linq;
-using System.Runtime.CompilerServices;
 using UnityEditor;
-using UnityEditor.Android;
-using System.Text;
 
 
 namespace Unity.Android.Logcat
@@ -32,6 +23,9 @@ namespace Unity.Android.Logcat
 
         private void Initialize()
         {
+            if (!AndroidBridge.AndroidExtensionsInstalled)
+                return;
+
             if (m_Runtime != null)
                 return;
 
@@ -49,4 +43,3 @@ namespace Unity.Android.Logcat
         }
     }
 }
-#endif