+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatSettingsProvider.cs
@@ -1,15 +1,6 @@
-#if PLATFORM_ANDROID
-using System.Collections.Generic;
-using System.Diagnostics;
 using System;
-using System.Text.RegularExpressions;
-using System.Linq;
-using System.Runtime.CompilerServices;
 using UnityEditor;
 using UnityEngine;
-using UnityEditor.Android;
-using System.Text;
-using UnityEngine.UIElements;
 
 namespace Unity.Android.Logcat
 {
@@ -76,11 +67,12 @@ namespace Unity.Android.Logcat
         }
 
         [SettingsProvider]
-        public static SettingsProvider CreateMyCustomSettingsProvider()
+        public static SettingsProvider CreateAndroidLogcatSettingsProvider()
         {
+            if (!AndroidBridge.AndroidExtensionsInstalled)
+                return null;
             var provider = new AndroidLogcatSettingsProvider(kSettingsPath, SettingsScope.User);
             return provider;
         }
     }
 }
-#endif