+++ com.unity.mobile.android-logcat/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2021-01-01
+
+### Fix & Improvements.
+ - Removed limitation where Android Logcat could only be used while active Editor platform is Android. Note: Android Support is still required to be installed for Android Logcat to work.
+
 ## [1.2.0] - 2020-08-18
 
 ### Fix & Improvements.